Audit.WebApi - Use GetCurrentAuditScope when an HttpContext or HttpRequest don't exist, such as in unit tests
Existing GetCurrentAuditScope assumes an HttpContext exists but this is not always the case, for example when the Controller classes are unit tested. In this case, a null reference is thrown.

One idea is to return a NoOp AuditScope to avoid producing side effects, like saving the events on the file system.